UBUNTU: SAUCE: apparmor4.0.0 [27/76]: Stacking v38: Fix prctl() syscall with apparmor=0
This should not be forwarded upstream as it fixes an issue in the currently
out of tree LSM stacking patchset that Ubuntu carries.

When apparmor=0 is specified as a boot option, the LSM no longer has a
"major" LSM registered and the condition lsm_slot == 0 will be true.

The lsm_slot == 0 check is present to protect against access to the
PR_LSM_ATTR_SET and PR_LSM_ATTR_GET command options being run without
a major LSM present but the check is too early preventing the regular
LSM hook processing (lsm_slot == 0 does not means there are not
minor lsms present).

Move the check after regular LSM hook processing to restrore the
previous behavior.

BugLink: http://bugs.launchpad.net/bugs/2016908
Fixes: c2350a7eca5c ("UBUNTU: SAUCE: Stacking v38: LSM: Specify which LSM to display")
Signed-off-by: John Johansen <john.johansen@canonical.com>