UBUNTU: SAUCE: apparmor4.0.0 [58/76]: prompt - ref count pdb
With the move to permission tables the dfa is no longer a stand
alone entity when used, needing a minimum of a permission table.
However it still could be shared among different pdbs each using
a different permission table.

Instead of duping the permission table when sharing a pdb, add a
refcount to the pdb so it can be easily shared.

Signed-off-by: John Johansen <john.johansen@canonical.com>