UBUNTU: SAUCE: apparmor4.0.0 [67/82]: prompt - fix notification cache when updating
BugLink: https://bugs.launchpad.net/bugs/2028253

The notification cache when updating doesn't take into account allowed
permissions that did not go to the user upcall as part of a deny.
If userspace only replies to the deny portion of the notification
then the allowed permissions won't be added to the cache.

Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from https://gitlab.com/jjohansen/apparmor-kernel)
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>