solus: Don't automatically compress modules
If CONFIG_MODULE_COMPRESS_ZSTD is set then the kernel will automatically compress modules when `make modules_install` is called. This is not ideal as we would prefer to handle uncompressed modules and would need to add an additional decompression step only to re-compress them later.

However, if CONFIG_MODULE_COMPRESS_ZSTD is not set then the kernel is not compiled with built-in module decompression support. This is less efficient for our purposes too since we'd rather the kernel do module decompression rather than delegating it to user-space.

So instead let's just force the `modules_install` target to skip compression entirely. This keeps kernel decompression support and allows us to work with uncompressed kernel modules