platform/x86: asus-wmi: store a min default for ppt options
Laptops with any of the ppt or nv tunables default to the minimum setting
on boot so we can safely assume a stored value is correct.

This patch adds storing of those values in the local struct, and enables
reading of those values back.

Secondary to the above it renames some internal variables to be more
consistent (which makes code grepping show all related parts)

Signed-off-by: Luke D. Jones <luke@ljones.dev>